Amélioration gestion Exceptions

+++ escalade/escalade-business/src/main/java/org/escalade/business/impl/manager/TopoManagerImpl.java
@@ -10,11 +10,14 @@ import org.apache.logging.log4j.Logger;
 import org.escalade.business.contract.manager.TopoManager;
 import org.escalade.model.bean.topo.Topo;
 import org.escalade.model.exception.FunctionalException;
+import org.escalade.model.exception.NotFoundException;
+import org.escalade.model.exception.TechnicalException;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.DefaultTransactionDefinition;
 
 /**
  * Implémnetation de TopoManager
+ * 
  * @author Oltenos
  *
  */
@@ -23,109 +26,107 @@ public class TopoManagerImpl extends AbstractManagerImpl implements TopoManager 
 	private static final Logger LOGGER = LogManager.getLogger(TopoManagerImpl.class);
 
 	@Override
-	public Topo getTopo(String titre) {
+	public Topo getTopo(String titre) throws NotFoundException, FunctionalException {
 		LOGGER.traceEntry("titre = " + titre);
 		
-		if(titre!=null) {
-			Topo result = this.getDaoFactory().getTopoDao().getTopo(titre);
-			
-			LOGGER.traceExit(result);
-			return result;
+		if (titre == null) {
+			throw new FunctionalException("Invalid informations sent to database");
 		}
 		
-		LOGGER.traceExit(null);
-		return null;
+		Topo result = this.getDaoFactory().getTopoDao().getTopo(titre);
+		
+		if (result == null) {
+			throw new NotFoundException("The book was not found with title = " + titre);
+		}
+
+		LOGGER.traceExit(result);
+		return result;
 	}
 
 	@Override
-	public void createTopo(Topo topo) throws FunctionalException {
+	public void createTopo(Topo topo) throws FunctionalException, TechnicalException {
 		LOGGER.traceEntry("topo = " + topo);
 		
+		if (topo == null) {
+			throw new FunctionalException("Invalid informations sent to database");
+		}
+
 		Set<ConstraintViolation<Topo>> violations = this.getValidator().validate(topo);
 		LOGGER.debug("resultat validation topo = " + violations);
-		
-		if(violations.isEmpty()) {
+
+		if (violations.isEmpty()) {
 			TransactionStatus vTransactionStatus = this.getPlatformTransactionManager().getTransaction(new DefaultTransactionDefinition());
 			try {
 				this.getDaoFactory().getTopoDao().createTopo(topo);
-	
+
 				TransactionStatus vTScommit = vTransactionStatus;
 				vTransactionStatus = null;
 				this.getPlatformTransactionManager().commit(vTScommit);
 			} finally {
 				if (vTransactionStatus != null) {
 					this.getPlatformTransactionManager().rollback(vTransactionStatus);
+					throw new TechnicalException("Technical error with the database");
 				}
 			}
-		}else {
-			String message = "";
-			int i=0;
-			for (ConstraintViolation<Topo> constraintViolation : violations) {
-				if(i==0) {
-					message += constraintViolation.getMessage();
-					i++;
-				}else{
-					message += ", " + constraintViolation.getMessage();
-				}
-			}
-			throw new FunctionalException(message);
+		} else {
+			throw new FunctionalException("Invalid informations sent to database");
 		}
-		
+
 		LOGGER.traceExit();
 	}
 
 	@Override
-	public void updateTopo(Topo topo) throws FunctionalException {
+	public void updateTopo(Topo topo) throws FunctionalException, TechnicalException {
 		LOGGER.traceEntry("topo = " + topo);
 		
+		if (topo == null) {
+			throw new FunctionalException("Invalid informations sent to database");
+		}
+
 		Set<ConstraintViolation<Topo>> violations = this.getValidator().validate(topo);
 		LOGGER.debug("resultat validation topo = " + violations);
-		
-		if(violations.isEmpty()) {
+
+		if (violations.isEmpty()) {
 			TransactionStatus vTransactionStatus = this.getPlatformTransactionManager().getTransaction(new DefaultTransactionDefinition());
 			try {
 				this.getDaoFactory().getTopoDao().updateTopo(topo);
-	
+
 				TransactionStatus vTScommit = vTransactionStatus;
 				vTransactionStatus = null;
 				this.getPlatformTransactionManager().commit(vTScommit);
 			} finally {
 				if (vTransactionStatus != null) {
 					this.getPlatformTransactionManager().rollback(vTransactionStatus);
+					throw new TechnicalException("Technical error with the database");
 				}
 			}
-		}else {
-			String message = "";
-			int i=0;
-			for (ConstraintViolation<Topo> constraintViolation : violations) {
-				if(i==0) {
-					message += constraintViolation.getMessage();
-					i++;
-				}else{
-					message += ", " + constraintViolation.getMessage();
-				}
-			}
-			throw new FunctionalException(message);
+		} else {
+			throw new FunctionalException("Invalid informations sent to database");
 		}
-		
+
 		LOGGER.traceExit();
 	}
 
 	@Override
-	public void deleteTopo(String titre) {
+	public void deleteTopo(String titre) throws  FunctionalException, TechnicalException  {
 		LOGGER.traceEntry("titre = " + titre);
 		
-		if(titre!=null) {
+		if (titre == null) {
+			throw new FunctionalException("Invalid informations sent to database");
+		}
+
+		if (titre != null) {
 			TransactionStatus vTransactionStatus = this.getPlatformTransactionManager().getTransaction(new DefaultTransactionDefinition());
 			try {
 				this.getDaoFactory().getTopoDao().deleteTopo(titre);
-	
+
 				TransactionStatus vTScommit = vTransactionStatus;
 				vTransactionStatus = null;
 				this.getPlatformTransactionManager().commit(vTScommit);
 			} finally {
 				if (vTransactionStatus != null) {
 					this.getPlatformTransactionManager().rollback(vTransactionStatus);
+					throw new TechnicalException("Technical error with the database");
 				}
 			}
 		}
